+++ src/components/layout/LayoutSidenav.vue
@@ -28,7 +28,7 @@
   </aside>
 </template>
 <script>
-import SidenavList from "./SidenavList.vue";
+import SidenavList from "@/components/layout/side/SidenavList.vue";
 import logo from "@/assets/img/logo-ct.png";
 
 export default {
