+++ src/components/layout/UserFooter.vue
